Adjust LoginViewModelTests
