--- Dokan/DokanTests/DokanTests.swift
@@ -1,48 +1,0 @@
-//
-//  DokanTests.swift
-//  DokanTests
-//
-//  Created by Ahmed M. Hassan on 10/06/2022.
-//
-
-import XCTest
-@testable import Dokan
-
-
-class DokanTests: XCTestCase {
-
-    // MARK: Properties
-    
-    var sut: LoginViewModel! // System Under Test
-    override func setUpWithError() throws {
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-        sut = LoginViewModel()
-    }
-
-    override func tearDownWithError() throws {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        sut = nil
-
-    }
-
-    func testConfigureOnButtonEnabled_WhenEmailIsEmpty_ShouldBeDisabled() {
-        // Given
-                var isEnabled: Bool?
-                sut.configureOnButtonEnabled { isEnabled = $0 }
-                // When
-                sut.updateEmail("")
-                // Then
-                XCTAssertTrue(isEnabled == false)
-                XCTAssertNotNil(isEnabled)
-    }
-    func testConfigureObButtonEnabled_WhenPasswordIsEmpty_ShouldBeDisabled() {
-           var isEnabled: Bool?
-           sut.configureOnButtonEnabled { isEnabled = $0 }
-           // When
-           sut.updatePassword("")
-           // Then
-           XCTAssertTrue(isEnabled == false)
-           XCTAssertNotNil(isEnabled)
-       }
-
-}
